+++ webpack.mix.js
@@ -15,4 +15,19 @@ mix.js('resources/assets/js/app.js', 'public/js')
    .sass('resources/assets/sass/app.scss', 'public/css')
     .options({
         processCssUrls: false
-    });
+    })
+    .browserSync(({
+        host: '192.168.10.10',
+        proxy: 'simgoodies.test',
+        open: false,
+        files: [
+            'app/**/*.php',
+            'resources/views/**/*.php',
+            'public/js/**/*.js',
+            'public/css/**/*.css'
+        ],
+        watchOptions: {
+            usePolling: true,
+            interval: 500
+        }
+    }));